Fix authentication

+++ backend/src/modules/business/business.entity.ts
@@ -35,7 +35,7 @@ export class BusinessEntity extends BaseEntityApp {
   @Column()
   currencyId: number;
 
-  @Column({ type: "money" })
+  @Column({ type: "float" })
   amount: number;
 
   @Column({
